+++ Vendedor.java
@@ -15,13 +15,13 @@ public class Vendedor extends Persona{
     }
     
     //comprar
-    public String comprar(String fecha,Proveedor pro,Producto p,float cant, Inventario i,ListaCompras lc){
+    public String comprar(String fecha,Proveedor pro,Producto p,double cant, Inventario i,ListaCompras lc){
         Compra comp =new Compra(fecha,clave,pro.getClave(),p.getClave(),cant,i);
         lc.anadirCompra(comp);
         return comp.toString();
     }
     //vender
-    public String vender(String fecha,Cliente cli,Producto p,float cant, Inventario i,VentasRealizadas vr){
+    public String vender(String fecha,Cliente cli,Producto p,double cant, Inventario i,VentasRealizadas vr){
         Venta vent =new Venta(fecha,clave,cli.getClave(),p.getClave(),cant,i);
         vr.anadirVenta(vent);
         return vent.toString();
